Added scripting dialog (very rudimentary).
Action log now also contains the set of selected elements before
each command. This allows to replay mesh generation scripts.

+++ src/tools/tool_manager.cpp
@@ -147,7 +147,7 @@ void ToolManager::execute_shortcut(int key, uint modifiers)
 		if((m_registeredTools[i].m_shortcutKey == key)
 		   && (m_registeredTools[i].m_shortcutModifiers == modifiers))
 		{
-			launchTool(i);
+			launchTool((int)i);
 			break;
 		}
 	}
